# packages/homeflow/openapi.yaml
openapi: 3.1.0
info:
  title: HomeFlow — Extropy Engine IoT Smart Home Service
  version: 1.0.0
  description: |
    HomeFlow is an IoT-integrated smart home/building management service
    built on top of the Extropy Engine. It measures thermodynamic entropy
    reduction from home automation actions and mints XP through the
    verification loop.

    Core formula: XP = R × F × ΔS × (w · E) × log(1/Tₛ)

    HomeFlow operates in the THERMODYNAMIC entropy domain and interoperates
    with all 13+ ecosystem apps via the shared EventPayloadMap and a
    plugin/adapter interoperability layer.
  contact:
    name: Extropy Engine
    url: https://github.com/00ranman/extropy-engine
  license:
    name: MIT
servers:
  - url: http://localhost:4015
    description: Local development
  - url: http://homeflow:4015
    description: Docker Compose internal
...
paths:
...
  /api/v1/households:
    post:
      tags: [Households]
      summary: Create a household
      operationId: createHousehold
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateHouseholdRequest'
      responses:
        '201':
          description: Household created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Household'
  /api/v1/households/{id}:
    get:
      tags: [Households]
      summary: Get household details
      operationId: getHousehold
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Household details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Household'
        '404':
          description: Household not found
    patch:
      tags: [Households]
      summary: Update household
      operationId: updateHousehold
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Updated household
    delete:
      tags: [Households]
      summary: Delete household
      operationId: deleteHousehold
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted
  /api/v1/zones:
    post:
      tags: [Zones]
      summary: Create a zone
      operationId: createZone
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateZoneRequest'
      responses:
        '201':
          description: Zone created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Zone'
    get:
      tags: [Zones]
      summary: List zones for household
      operationId: listZones
      parameters:
        - name: householdId
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Zones list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedZones'
  /api/v1/devices:
    post:
      tags: [Devices]
      summary: Register a new IoT device
      operationId: registerDevice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDeviceRequest'
      responses:
        '201':
          description: Device registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
    get:
      tags: [Devices]
      summary: List devices for household
      operationId: listDevices
      parameters:
        - name: householdId
          in: query
          required: true
          schema:
            type: string
        - name: type
          in: query
          schema:
            $ref: '#/components/schemas/DeviceType'
      responses:
        '200':
          description: Devices list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedDevices'
  /api/v1/devices/{id}:
    get:
      tags: [Devices]
      summary: Get device details
      operationId: getDevice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Device details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '404':
          description: Device not found
    patch:
      tags: [Devices]
      summary: Update device
      operationId: updateDevice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Updated device
    delete:
      tags: [Devices]
      summary: Delete device
      operationId: deleteDevice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Deleted
  /api/v1/devices/{id}/commands:
    post:
      tags: [Devices]
      summary: Issue a command to a device
      description: |
        Send a command to an IoT device (e.g., set thermostat to 72°F,
        turn off lights, change HVAC mode). The command is confirmed
        immediately in dev mode.
      operationId: issueCommand
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IssueCommandRequest'
            examples:
              setThermostat:
                summary: Set thermostat to 72°F
                value:
                  commandType: set_temperature
                  parameters:
                    temperatureF: 72
                  issuedBy: validator-001
              toggleLight:
                summary: Turn off lights
                value:
                  commandType: toggle_power
                  parameters:
                    on: false
                  issuedBy: validator-001
      responses:
        '201':
          description: Command issued and confirmed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceCommand'
    get:
      tags: [Devices]
      summary: Get command history for device
      operationId: getCommandHistory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Command history
  /api/v1/entropy/snapshot:
    post:
      tags: [Entropy]
      summary: Take an entropy snapshot
      description: |
        Reads all device states for a household and computes aggregate
        thermodynamic entropy (S = Q/T in J/K).
      operationId: takeSnapshot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [householdId]
              properties:
                householdId:
                  type: string
      responses:
        '201':
          description: Snapshot taken
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HomeEntropySnapshot'
  /api/v1/entropy/measure:
    post:
      tags: [Entropy]
      summary: Measure entropy reduction and auto-generate claim
      description: |
        Compares the latest two entropy snapshots. If ΔS > 0, automatically
        generates a claim to the Epistemology Engine and opens a loop in
        the Loop Ledger.

        ΔS breakdown:
        - Energy savings (kWh reduction → J/K)
        - Temperature optimization (deviation from setpoint)
        - Resource efficiency (solar utilization)
      operationId: measureReduction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [householdId]
              properties:
                householdId:
                  type: string
                causalCommandIds:
                  type: array
                  items:
                    type: string
      responses:
        '201':
          description: Entropy reduction measured and claim generated
          content:
            application/json:
              schema:
                type: object
                properties:
                  reduction:
                    $ref: '#/components/schemas/HomeEntropyReduction'
                  claim:
                    type: object
                    properties:
                      claimId:
                        type: string
                      loopId:
                        type: string
  /api/v1/entropy/{householdId}/history:
    get:
      tags: [Entropy]
      summary: Get entropy reduction history
      operationId: getEntropyHistory
      parameters:
        - name: householdId
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Entropy history
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/HomeEntropyReduction'
                  cumulativeDeltaS:
                    type: number
  /api/v1/entropy/claims/{householdId}:
    get:
      tags: [Entropy]
      summary: Get claims history for household
      operationId: getClaimsHistory
      parameters:
        - name: householdId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Claims history
...
components:
  schemas:
...
    DeviceType:
      type: string
      enum:
        - thermostat
        - temperature_sensor
        - humidity_sensor
        - lighting
        - hvac
        - energy_monitor
        - smart_plug
        - occupancy_sensor
        - water_meter
        - solar_panel
        - battery_storage
        - air_quality
    CommandType:
      type: string
      enum:
        - set_temperature
        - set_brightness
        - set_hvac_mode
        - set_fan_speed
        - toggle_power
        - set_schedule
        - trigger_scene
    Device:
      type: object
      properties:
        id:
          type: string
        householdId:
          type: string
        zoneId:
          type: string
          nullable: true
        name:
          type: string
        type:
          $ref: '#/components/schemas/DeviceType'
        manufacturer:
          type: string
        model:
          type: string
        firmwareVersion:
          type: string
        status:
          type: string
          enum: [online, offline, error, pairing, updating]
        capabilities:
          type: array
          items:
            type: object
        state:
          type: object
          additionalProperties: true
        metadata:
          type: object
        lastSeenAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Household:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        dfaoId:
          type: string
          nullable: true
        validatorId:
          type: string
        memberValidatorIds:
          type: array
          items:
            type: string
        address:
          type: string
          nullable: true
        timezone:
          type: string
        area_sqft:
          type: number
          nullable: true
        zoneIds:
          type: array
          items:
            type: string
        energyBaselineKwh:
          type: number
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Zone:
      type: object
      properties:
        id:
          type: string
        householdId:
          type: string
        name:
          type: string
        floor:
          type: integer
        area_sqft:
          type: integer
        deviceIds:
          type: array
          items:
            type: string
        targetTemperatureF:
          type: number
          nullable: true
        targetHumidityPct:
          type: number
          nullable: true
        isOccupied:
          type: boolean
    DeviceCommand:
      type: object
      properties:
        id:
          type: string
        deviceId:
          type: string
        type:
          $ref: '#/components/schemas/CommandType'
        parameters:
          type: object
        status:
          type: string
          enum: [pending, sent, confirmed, failed, timeout]
        issuedBy:
          type: string
        issuedAt:
          type: string
          format: date-time
        confirmedAt:
          type: string
          format: date-time
          nullable: true
        previousState:
          type: object
        newState:
          type: object
    HomeEntropySnapshot:
      type: object
      properties:
        householdId:
          type: string
        timestamp:
          type: string
          format: date-time
        totalPowerWatts:
          type: number
        energyConsumedWh:
          type: number
        avgIndoorTempF:
          type: number
        outdoorTempF:
          type: number
          nullable: true
        avgHumidityPct:
          type: number
        solarGeneratedWh:
          type: number
        occupiedZones:
          type: integer
        totalZones:
          type: integer
        entropyJoulePerKelvin:
          type: number
    HomeEntropyReduction:
      type: object
      properties:
        householdId:
          type: string
        deltaS:
          type: number
          description: 'Entropy reduction in J/K (must be > 0 for XP)'
        breakdown:
          type: object
          properties:
            energySavingsJK:
              type: number
            temperatureOptimizationJK:
              type: number
            resourceEfficiencyJK:
              type: number
        causalCommandIds:
          type: array
          items:
            type: string
        confidence:
          type: number
          minimum: 0
          maximum: 1
        measuredAt:
          type: string
          format: date-time
...
    RegisterDeviceRequest:
      type: object
      required: [householdId, name, type, manufacturer, model, firmwareVersion]
      properties:
        householdId:
          type: string
        name:
          type: string
        type:
          $ref: '#/components/schemas/DeviceType'
        manufacturer:
          type: string
        model:
          type: string
        firmwareVersion:
          type: string
        zoneId:
          type: string
        capabilities:
          type: array
          items:
            type: object
        metadata:
          type: object
    IssueCommandRequest:
      type: object
      required: [commandType, issuedBy]
      properties:
        commandType:
          $ref: '#/components/schemas/CommandType'
        parameters:
          type: object
        issuedBy:
          type: string
    CreateHouseholdRequest:
      type: object
      required: [name, validatorId]
      properties:
        name:
          type: string
        validatorId:
          type: string
        address:
          type: string
        timezone:
          type: string
          default: America/Chicago
        area_sqft:
          type: number
        energyBaselineKwh:
          type: number
    CreateZoneRequest:
      type: object
      required: [householdId, name, floor, area_sqft]
      properties:
        householdId:
          type: string
        name:
          type: string
        floor:
          type: integer
        area_sqft:
          type: integer
        targetTemperatureF:
          type: number
        targetHumidityPct:
          type: number
...
    PaginatedDevices:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Device'
        total:
          type: integer
    PaginatedZones:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Zone'
        total:
          type: integer